Move Metabot editor extensions and metabase:// link utils across the metabot/rich_text_editing boundary
- Move tiptap MetabotEmbed and MetabotMention extensions from
  metabase/rich_text_editing into metabase/metabot/components/editor-extensions;
  their imports of rich_text_editing internals become absolute imports in the
  kept metabot->rte direction.
- Move metabase:// protocol link utils from metabase/metabot/utils/links to
  metabase/rich_text_editing/utils/links, since rich_text_editing (SmartLink,
  suggestionUtils) must not import from metabase/metabot.
- Update all importers (documents Editor, PublicDocument, MetabotPromptInput,
  AIMarkdown, MetabotChainOfThought, specs) to the new paths.
