+++ frontend/src/metabase/metabot/components/MetabotPromptInput/MetabotPromptInput.unit.spec.tsx
@@ -14,8 +14,8 @@ import { mockSettings } from "__support__/settings";
 import { renderWithProviders } from "__support__/ui";
 import { ROOT_COLLECTION } from "metabase/common/collections/constants";
 import type { MetabotPromptInputRef } from "metabase/metabot";
+import { MetabotMentionPluginKey } from "metabase/metabot/components/editor-extensions/MetabotMention/MetabotMentionExtension";
 import { createMockState } from "metabase/redux/store/mocks";
-import { MetabotMentionPluginKey } from "metabase/rich_text_editing/tiptap/extensions/MetabotMention/MetabotMentionExtension";
 import type { SuggestionModel } from "metabase/rich_text_editing/tiptap/extensions/shared/types";
 import {
   createMockCollection,
