+++ frontend/src/metabase/metabot/components/editor-extensions/MetabotEmbed/MetabotEmbed.unit.spec.tsx
