+++ frontend/src/metabase/metabot/components/editor-extensions/MetabotEmbed/__support__/editor-mocks.ts
