+++ frontend/src/metabase/metabot/components/editor-extensions/MetabotEmbed/__support__/node-view-mocks.ts
