+++ frontend/src/metabase/metabot/components/editor-extensions/MetabotMention/MetabotSuggestion.tsx
@@ -19,10 +19,9 @@ import { useMetabotName } from "metabase/metabot/hooks";
 import type { MenuItem } from "metabase/rich_text_editing/tiptap/extensions/shared/MenuComponents";
 import { MenuItemComponent } from "metabase/rich_text_editing/tiptap/extensions/shared/MenuComponents";
 import { SuggestionPaper } from "metabase/rich_text_editing/tiptap/extensions/shared/SuggestionPaper";
+import { useBuildDbMenuItems } from "metabase/rich_text_editing/tiptap/extensions/shared/suggestionHooks";
 import { Box, Group, Loader, Text } from "metabase/ui";
 import type { Database, SearchResult } from "metabase-types/api";
-
-import { useBuildDbMenuItems } from "../shared/suggestionHooks";
 
 interface MentionSuggestionProps {
   items: SearchResult[];
